+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImplInternal.java
@@ -938,7 +938,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
           livePlaybackSpeedControl.getAdjustedPlaybackSpeed(
               getCurrentLiveOffsetUs(), getTotalBufferedDurationUs());
       if (mediaClock.getPlaybackParameters().speed != adjustedSpeed) {
-        mediaClock.setPlaybackParameters(playbackInfo.playbackParameters.withSpeed(adjustedSpeed));
+        setMediaClockPlaybackParameters(playbackInfo.playbackParameters.withSpeed(adjustedSpeed));
         handlePlaybackParameters(
             playbackInfo.playbackParameters,
             /* currentPlaybackSpeed= */ mediaClock.getPlaybackParameters().speed,
@@ -946,6 +946,12 @@ import java.util.concurrent.atomic.AtomicBoolean;
             /* acknowledgeCommand= */ false);
       }
     }
+  }
+
+  private void setMediaClockPlaybackParameters(PlaybackParameters playbackParameters) {
+    // Previously sent speed updates from the media clock now become stale.
+    handler.removeMessages(MSG_PLAYBACK_PARAMETERS_CHANGED_INTERNAL);
+    mediaClock.setPlaybackParameters(playbackParameters);
   }
 
   private void notifyTrackSelectionRebuffer() {
@@ -1334,7 +1340,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
   private void setPlaybackParametersInternal(PlaybackParameters playbackParameters)
       throws ExoPlaybackException {
-    mediaClock.setPlaybackParameters(playbackParameters);
+    setMediaClockPlaybackParameters(playbackParameters);
     handlePlaybackParameters(mediaClock.getPlaybackParameters(), /* acknowledgeCommand= */ true);
   }
 
@@ -1649,7 +1655,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
     nextPendingMessageIndexHint = nextPendingMessageIndex;
   }
 
-  private void ensureStopped(Renderer renderer) throws ExoPlaybackException {
+  private void ensureStopped(Renderer renderer) {
     if (renderer.getState() == Renderer.STATE_STARTED) {
       renderer.stop();
     }
@@ -1906,14 +1912,20 @@ import java.util.concurrent.atomic.AtomicBoolean;
       MediaPeriodId newPeriodId,
       Timeline oldTimeline,
       MediaPeriodId oldPeriodId,
-      long positionForTargetOffsetOverrideUs) {
+      long positionForTargetOffsetOverrideUs)
+      throws ExoPlaybackException {
     if (!shouldUseLivePlaybackSpeedControl(newTimeline, newPeriodId)) {
       // Live playback speed control is unused for the current period, reset speed to user-defined
       // playback parameters or 1.0 for ad playback.
       PlaybackParameters targetPlaybackParameters =
           newPeriodId.isAd() ? PlaybackParameters.DEFAULT : playbackInfo.playbackParameters;
       if (!mediaClock.getPlaybackParameters().equals(targetPlaybackParameters)) {
-        mediaClock.setPlaybackParameters(targetPlaybackParameters);
+        setMediaClockPlaybackParameters(targetPlaybackParameters);
+        handlePlaybackParameters(
+            playbackInfo.playbackParameters,
+            targetPlaybackParameters.speed,
+            /* updatePlaybackInfo= */ false,
+            /* acknowledgeCommand= */ false);
       }
       return;
     }
@@ -1962,7 +1974,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
     return maxReadPositionUs;
   }
 
-  private void updatePeriods() throws ExoPlaybackException, IOException {
+  private void updatePeriods() throws ExoPlaybackException {
     if (playbackInfo.timeline.isEmpty() || !mediaSourceList.isPrepared()) {
       // No periods available.
       return;
@@ -2004,7 +2016,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
     }
   }
 
-  private void maybeUpdateReadingPeriod() {
+  private void maybeUpdateReadingPeriod() throws ExoPlaybackException {
     @Nullable MediaPeriodHolder readingPeriodHolder = queue.getReadingPeriod();
     if (readingPeriodHolder == null) {
       return;