+++ library/core/src/main/java/com/google/android/exoplayer2/audio/AudioTrackPositionTracker.java
@@ -15,6 +15,7 @@
  */
 package com.google.android.exoplayer2.audio;
 
+import static com.google.android.exoplayer2.util.Assertions.checkNotNull;
 import static com.google.android.exoplayer2.util.Util.castNonNull;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
@@ -26,7 +27,6 @@ import android.os.SystemClock;
 import androidx.annotation.IntDef;
 import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Util;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
@@ -146,6 +146,9 @@ import java.lang.reflect.Method;
   /** The duration of time used to smooth over an adjustment between position sampling modes. */
   private static final long MODE_SWITCH_SMOOTHING_DURATION_US = C.MICROS_PER_SECOND;
 
+  /** Minimum update interval for getting the raw playback head position, in milliseconds. */
+  private static final long RAW_PLAYBACK_HEAD_POSITION_UPDATE_INTERVAL_MS = 5;
+
   private static final long FORCE_RESET_WORKAROUND_TIMEOUT_MS = 200;
 
   private static final int MAX_PLAYHEAD_OFFSET_COUNT = 10;
@@ -174,7 +177,8 @@ import java.lang.reflect.Method;
 
   private boolean isOutputPcm;
   private long lastLatencySampleTimeUs;
-  private long lastRawPlaybackHeadPosition;
+  private long lastRawPlaybackHeadPositionSampleTimeMs;
+  private long rawPlaybackHeadPosition;
   private long rawPlaybackHeadWrapCount;
   private long passthroughWorkaroundPauseOffset;
   private int nextPlayheadOffsetIndex;
@@ -199,7 +203,7 @@ import java.lang.reflect.Method;
    * @param listener A listener for position tracking events.
    */
   public AudioTrackPositionTracker(Listener listener) {
-    this.listener = Assertions.checkNotNull(listener);
+    this.listener = checkNotNull(listener);
     if (Util.SDK_INT >= 18) {
       try {
         getLatencyMethod = AudioTrack.class.getMethod("getLatency", (Class<?>[]) null);
@@ -235,7 +239,7 @@ import java.lang.reflect.Method;
     needsPassthroughWorkarounds = isPassthrough && needsPassthroughWorkarounds(outputEncoding);
     isOutputPcm = Util.isEncodingLinearPcm(outputEncoding);
     bufferSizeUs = isOutputPcm ? framesToDurationUs(bufferSize / outputPcmFrameSize) : C.TIME_UNSET;
-    lastRawPlaybackHeadPosition = 0;
+    rawPlaybackHeadPosition = 0;
     rawPlaybackHeadWrapCount = 0;
     passthroughWorkaroundPauseOffset = 0;
     hasData = false;
@@ -253,10 +257,11 @@ import java.lang.reflect.Method;
     if (audioTimestampPoller != null) {
       audioTimestampPoller.reset();
     }
+    resetSyncParams();
   }
 
   public long getCurrentPositionUs(boolean sourceEnded) {
-    if (Assertions.checkNotNull(this.audioTrack).getPlayState() == PLAYSTATE_PLAYING) {
+    if (checkNotNull(this.audioTrack).getPlayState() == PLAYSTATE_PLAYING) {
       maybeSampleSyncParams();
     }
 
@@ -264,7 +269,7 @@ import java.lang.reflect.Method;
     // Otherwise, derive a smoothed position by sampling the track's frame position.
     long systemTimeUs = System.nanoTime() / 1000;
     long positionUs;
-    AudioTimestampPoller audioTimestampPoller = Assertions.checkNotNull(this.audioTimestampPoller);
+    AudioTimestampPoller audioTimestampPoller = checkNotNull(this.audioTimestampPoller);
     boolean useGetTimestampMode = audioTimestampPoller.hasAdvancingTimestamp();
     if (useGetTimestampMode) {
       // Calculate the speed-adjusted position using the timestamp (which may be in the future).
@@ -282,7 +287,9 @@ import java.lang.reflect.Method;
         // getPlaybackHeadPositionUs() only has a granularity of ~20 ms, so we base the position off
         // the system clock (and a smoothed offset between it and the playhead position) so as to
         // prevent jitter in the reported positions.
-        positionUs = systemTimeUs + smoothedPlayheadOffsetUs;
+        positionUs =
+            Util.getMediaDurationForPlayoutDuration(
+                systemTimeUs + smoothedPlayheadOffsetUs, audioTrackPlaybackSpeed);
       }
       if (!sourceEnded) {
         positionUs = max(0, positionUs - latencyUs);
@@ -329,12 +336,12 @@ import java.lang.reflect.Method;
 
   /** Starts position tracking. Must be called immediately before {@link AudioTrack#play()}. */
   public void start() {
-    Assertions.checkNotNull(audioTimestampPoller).reset();
+    checkNotNull(audioTimestampPoller).reset();
   }
 
   /** Returns whether the audio track is in the playing state. */
   public boolean isPlaying() {
-    return Assertions.checkNotNull(audioTrack).getPlayState() == PLAYSTATE_PLAYING;
+    return checkNotNull(audioTrack).getPlayState() == PLAYSTATE_PLAYING;
   }
 
   /**
@@ -345,7 +352,7 @@ import java.lang.reflect.Method;
    * @return Whether the caller can write data to the track.
    */
   public boolean mayHandleBuffer(long writtenFrames) {
-    @PlayState int playState = Assertions.checkNotNull(audioTrack).getPlayState();
+    @PlayState int playState = checkNotNull(audioTrack).getPlayState();
     if (needsPassthroughWorkarounds) {
       // An AC-3 audio track continues to play data written while it is paused. Stop writing so its
       // buffer empties. See [Internal: b/18899620].
@@ -426,7 +433,7 @@ import java.lang.reflect.Method;
     if (stopTimestampUs == C.TIME_UNSET) {
       // The audio track is going to be paused, so reset the timestamp poller to ensure it doesn't
       // supply an advancing position.
-      Assertions.checkNotNull(audioTimestampPoller).reset();
+      checkNotNull(audioTimestampPoller).reset();
       return true;
     }
     // We've handled the end of the stream already, so there's no need to pause the track.
@@ -444,15 +451,17 @@ import java.lang.reflect.Method;
   }
 
   private void maybeSampleSyncParams() {
-    long playbackPositionUs = getPlaybackHeadPositionUs();
-    if (playbackPositionUs == 0) {
-      // The AudioTrack hasn't output anything yet.
-      return;
-    }
     long systemTimeUs = System.nanoTime() / 1000;
     if (systemTimeUs - lastPlayheadSampleTimeUs >= MIN_PLAYHEAD_OFFSET_SAMPLE_INTERVAL_US) {
+      long playbackPositionUs = getPlaybackHeadPositionUs();
+      if (playbackPositionUs == 0) {
+        // The AudioTrack hasn't output anything yet.
+        return;
+      }
       // Take a new sample and update the smoothed offset between the system clock and the playhead.
-      playheadOffsets[nextPlayheadOffsetIndex] = playbackPositionUs - systemTimeUs;
+      playheadOffsets[nextPlayheadOffsetIndex] =
+          Util.getPlayoutDurationForMediaDuration(playbackPositionUs, audioTrackPlaybackSpeed)
+              - systemTimeUs;
       nextPlayheadOffsetIndex = (nextPlayheadOffsetIndex + 1) % MAX_PLAYHEAD_OFFSET_COUNT;
       if (playheadOffsetCount < MAX_PLAYHEAD_OFFSET_COUNT) {
         playheadOffsetCount++;
@@ -470,12 +479,12 @@ import java.lang.reflect.Method;
       return;
     }
 
-    maybePollAndCheckTimestamp(systemTimeUs, playbackPositionUs);
+    maybePollAndCheckTimestamp(systemTimeUs);
     maybeUpdateLatency(systemTimeUs);
   }
 
-  private void maybePollAndCheckTimestamp(long systemTimeUs, long playbackPositionUs) {
-    AudioTimestampPoller audioTimestampPoller = Assertions.checkNotNull(this.audioTimestampPoller);
+  private void maybePollAndCheckTimestamp(long systemTimeUs) {
+    AudioTimestampPoller audioTimestampPoller = checkNotNull(this.audioTimestampPoller);
     if (!audioTimestampPoller.maybePollTimestamp(systemTimeUs)) {
       return;
     }
@@ -483,6 +492,7 @@ import java.lang.reflect.Method;
     // Check the timestamp and accept/reject it.
     long audioTimestampSystemTimeUs = audioTimestampPoller.getTimestampSystemTimeUs();
     long audioTimestampPositionFrames = audioTimestampPoller.getTimestampPositionFrames();
+    long playbackPositionUs = getPlaybackHeadPositionUs();
     if (Math.abs(audioTimestampSystemTimeUs - systemTimeUs) > MAX_AUDIO_TIMESTAMP_OFFSET_US) {
       listener.onSystemTimeUsMismatch(
           audioTimestampPositionFrames,
@@ -511,8 +521,7 @@ import java.lang.reflect.Method;
         // Compute the audio track latency, excluding the latency due to the buffer (leaving
         // latency due to the mixer and audio hardware driver).
         latencyUs =
-            castNonNull((Integer) getLatencyMethod.invoke(Assertions.checkNotNull(audioTrack)))
-                    * 1000L
+            castNonNull((Integer) getLatencyMethod.invoke(checkNotNull(audioTrack))) * 1000L
                 - bufferSizeUs;
         // Check that the latency is non-negative.
         latencyUs = max(latencyUs, 0);
@@ -550,7 +559,7 @@ import java.lang.reflect.Method;
    */
   private boolean forceHasPendingData() {
     return needsPassthroughWorkarounds
-        && Assertions.checkNotNull(audioTrack).getPlayState() == AudioTrack.PLAYSTATE_PAUSED
+        && checkNotNull(audioTrack).getPlayState() == AudioTrack.PLAYSTATE_PAUSED
         && getPlaybackHeadPosition() == 0;
   }
 
@@ -576,34 +585,44 @@ import java.lang.reflect.Method;
    * @return The playback head position, in frames.
    */
   private long getPlaybackHeadPosition() {
-    AudioTrack audioTrack = Assertions.checkNotNull(this.audioTrack);
+    long currentTimeMs = SystemClock.elapsedRealtime();
     if (stopTimestampUs != C.TIME_UNSET) {
       // Simulate the playback head position up to the total number of frames submitted.
-      long elapsedTimeSinceStopUs = (SystemClock.elapsedRealtime() * 1000) - stopTimestampUs;
-      long framesSinceStop = (elapsedTimeSinceStopUs * outputSampleRate) / C.MICROS_PER_SECOND;
+      long elapsedTimeSinceStopUs = (currentTimeMs * 1000) - stopTimestampUs;
+      long mediaTimeSinceStopUs =
+          Util.getMediaDurationForPlayoutDuration(elapsedTimeSinceStopUs, audioTrackPlaybackSpeed);
+      long framesSinceStop = (mediaTimeSinceStopUs * outputSampleRate) / C.MICROS_PER_SECOND;
       return min(endPlaybackHeadPosition, stopPlaybackHeadPosition + framesSinceStop);
     }
+    if (currentTimeMs - lastRawPlaybackHeadPositionSampleTimeMs
+        >= RAW_PLAYBACK_HEAD_POSITION_UPDATE_INTERVAL_MS) {
+      updateRawPlaybackHeadPosition(currentTimeMs);
+      lastRawPlaybackHeadPositionSampleTimeMs = currentTimeMs;
+    }
+    return rawPlaybackHeadPosition + (rawPlaybackHeadWrapCount << 32);
+  }
 
+  private void updateRawPlaybackHeadPosition(long currentTimeMs) {
+    AudioTrack audioTrack = checkNotNull(this.audioTrack);
     int state = audioTrack.getPlayState();
     if (state == PLAYSTATE_STOPPED) {
-      // The audio track hasn't been started.
-      return 0;
+      // The audio track hasn't been started. Keep initial zero timestamp.
+      return;
     }
-
     long rawPlaybackHeadPosition = 0xFFFFFFFFL & audioTrack.getPlaybackHeadPosition();
     if (needsPassthroughWorkarounds) {
       // Work around an issue with passthrough/direct AudioTracks on platform API versions 21/22
       // where the playback head position jumps back to zero on paused passthrough/direct audio
       // tracks. See [Internal: b/19187573].
       if (state == PLAYSTATE_PAUSED && rawPlaybackHeadPosition == 0) {
-        passthroughWorkaroundPauseOffset = lastRawPlaybackHeadPosition;
+        passthroughWorkaroundPauseOffset = this.rawPlaybackHeadPosition;
       }
       rawPlaybackHeadPosition += passthroughWorkaroundPauseOffset;
     }
 
     if (Util.SDK_INT <= 29) {
       if (rawPlaybackHeadPosition == 0
-          && lastRawPlaybackHeadPosition > 0
+          && this.rawPlaybackHeadPosition > 0
           && state == PLAYSTATE_PLAYING) {
         // If connecting a Bluetooth audio device fails, the AudioTrack may be left in a state
         // where its Java API is in the playing state, but the native track is stopped. When this
@@ -611,19 +630,18 @@ import java.lang.reflect.Method;
         // playback head position and force the track to be reset after
         // {@link #FORCE_RESET_WORKAROUND_TIMEOUT_MS} has elapsed.
         if (forceResetWorkaroundTimeMs == C.TIME_UNSET) {
-          forceResetWorkaroundTimeMs = SystemClock.elapsedRealtime();
+          forceResetWorkaroundTimeMs = currentTimeMs;
         }
-        return lastRawPlaybackHeadPosition;
+        return;
       } else {
         forceResetWorkaroundTimeMs = C.TIME_UNSET;
       }
     }
 
-    if (lastRawPlaybackHeadPosition > rawPlaybackHeadPosition) {
+    if (this.rawPlaybackHeadPosition > rawPlaybackHeadPosition) {
       // The value must have wrapped around.
       rawPlaybackHeadWrapCount++;
     }
-    lastRawPlaybackHeadPosition = rawPlaybackHeadPosition;
-    return rawPlaybackHeadPosition + (rawPlaybackHeadWrapCount << 32);
+    this.rawPlaybackHeadPosition = rawPlaybackHeadPosition;
   }
 }