+++ library/extractor/src/main/java/com/google/android/exoplayer2/text/webvtt/WebvttDecoder.java
@@ -28,7 +28,7 @@ import java.util.List;
 /**
  * A {@link SimpleSubtitleDecoder} for WebVTT.
  *
- * @see <a href="http://dev.w3.org/html5/webvtt">WebVTT specification</a>
+ * <p>See the <a href="http://dev.w3.org/html5/webvtt">WebVTT specification</a>.
  */
 public final class WebvttDecoder extends SimpleSubtitleDecoder {