+++ library/smoothstreaming/src/main/java/com/google/android/exoplayer2/source/smoothstreaming/manifest/SsManifest.java
@@ -33,8 +33,8 @@ import java.util.UUID;
 /**
  * Represents a SmoothStreaming manifest.
  *
- * @see <a href="http://msdn.microsoft.com/en-us/library/ee673436(v=vs.90).aspx">IIS Smooth
- *     Streaming Client Manifest Format</a>
+ * <p>See the <a href="http://msdn.microsoft.com/en-us/library/ee673436(v=vs.90).aspx">IIS Smooth
+ * Streaming Client Manifest Format</a>
  */
 public class SsManifest implements FilterableManifest<SsManifest> {